+++ app/controllers/node_info_controller.rb
@@ -1,7 +1,4 @@
 class NodeInfoController < ApplicationController
-  respond_to :json
-  respond_to :html, only: :statistics
-
   def jrd
     render json: NodeInfo.jrd(CGI.unescape(node_info_url("123.123").sub("123.123", "%{version}")))
   end
@@ -16,9 +13,6 @@ class NodeInfoController < ApplicationController
   end
 
   def statistics
-    respond_to do |format|
-      format.json { render json: StatisticsPresenter.new }
-      format.all { @statistics = NodeInfoPresenter.new("1.0") }
-    end
+    @statistics = NodeInfoPresenter.new("1.0")
   end
 end