+++ spec/controllers/node_info_controller_spec.rb
@@ -51,17 +51,6 @@ describe NodeInfoController do
   end
 
   describe "#statistics" do
-    it "responds to format json" do
-      get :statistics, format: "json"
-      expect(response.code).to eq("200")
-    end
-
-    it "contains json" do
-      get :statistics, format: "json"
-      json = JSON.parse(response.body)
-      expect(json["name"]).to be_present
-    end
-
     it "responds to html" do
       get :statistics, format: "html"
       expect(response.code).to eq("200")